requirements.txt
streamlit
openai
python-dotenv
beautifulsoup4
pandas
scikit-learn
nltk
